refactor(email): switch from SMTP to SendGrid for sending emails
