+++ requirements.txt
@@ -11,3 +11,4 @@ python-multipart==0.0.6
 pytest==7.4.3
 pytest-asyncio==0.21.1
 httpx==0.25.2
+sendgrid==6.12.5
